+++ include/staff/system.inc.php
@@ -49,7 +49,35 @@ $extensions = array(
 <tbody>
     <tr><td><?php echo __('osTicket Version'); ?></td>
         <td><span class="ltr"><?php
-        echo sprintf("%s (%s)", THIS_VERSION, trim($commit)); ?></span></td></tr>
+            echo sprintf("%s (%s)", THIS_VERSION, trim($commit)); ?></span>
+<?php
+$lv = $ost->getLatestVersion('core', MAJOR_VERSION);
+$tv = THIS_VERSION;
+$gv = GIT_VERSION == '$git' ? substr(@`git rev-parse HEAD`, 0, 7) : false ?: GIT_VERSION;
+if ($lv && $tv[0] == 'v' ? version_compare(THIS_VERSION, $lv, '>=') : $lv == $gv) { ?>
+    — <span style="color:green"><i class="icon-check"></i> <?php echo __('Up to date'); ?></span>
+<?php
+}
+else {
+    // Report current version (v1.9.x ?: deadbeef ?: $git)
+    $cv = $tv[0] == 'v' ? $tv : $gv;
+?>
+      <a class="green button action-button pull-right"
+         href="http://osticket.com/download?cv=<?php echo $cv; ?>"><i class="icon-rocket"></i>
+        <?php echo __('Upgrade'); ?></a>
+<?php if ($lv) { ?>
+      <strong> — <?php echo str_replace(
+          '%s', $lv, __("%s is available")
+      ); ?></strong>
+<?php }
+}
+if (!$lv) { ?>
+    <strong> — <?php echo __('This osTicket version is no longer supported. Please consider upgrading');
+        ?></strong>
+<?php
+}
+?>
+    </td></tr>
     <tr><td><?php echo __('Web Server Software'); ?></td>
         <td><span class="ltr"><?php echo $_SERVER['SERVER_SOFTWARE']; ?></span></td></tr>
     <tr><td><?php echo __('MySQL Version'); ?></td>